requirements.txt
apturl==0.5.2
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
auth==0.5.3
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
duplicity==0.8.12.0
entrypoints==0.3
eventlet==0.33.1
falcon==3.1.0
fasteners==0.14.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupSafe==1.1.0
mongoengine==0.24.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
outcome==1.2.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
segno==1.5.2
selenium==4.2.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
systemd-python==234
trio==0.21.0
trio-websocket==0.9.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==2.1.2
whitenoise==6.2.0
wsproto==1.1.0
xkit==0.0.0
zipp==3.8.0
